requirements.txt
discord.py
python-dotenv
sourceserver
pytz